requirements.txt
appnope==0.1.0
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
influxdb==5.3.0
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
wrapt==1.12.1
